+++ client/src/utils/colors.ts
@@ -1,0 +1,3 @@
+export const primaryColor = 'rgb(0, 255, 127)';
+export const primaryColor50 = 'rgb(0, 255, 127, 0.5)';
+export const secondaryColor = 'rgb(255, 75, 145)';
